Strife now boots
The problem was the opendir implementation.  The handle that is passed
around to _findnext and friends is supposed to be a intptr_t, but was
only a long instead.  I suspect the problem only showed up on 64-bit
builds beforehand.

+++ win32/win_opendir.h
@@ -46,7 +46,7 @@ typedef struct
    struct dirent dd_dir;
 
    /* _findnext handle */
-   long	dd_handle;
+   intptr_t	dd_handle;
 
    /*
     * Status of search:
